jubjub
Function batch_normalize

pub fn batch_normalize(
    v: &mut [ExtendedPoint],
) -> impl Iterator<Item = AffinePoint> + '_

This takes a mutable slice of ExtendedPoints and “normalizes” them using only a single inversion for the entire batch. This normalization results in all of the points having a Z-coordinate of one. Further, an iterator is returned which can be used to obtain AffinePoints for each element in the slice.

This costs 5 multiplications per element, and a field inversion.