jubjub
Struct SubgroupPoint

pub struct SubgroupPoint(/* private fields */);

This represents a point in the prime-order subgroup of Jubjub, in extended coordinates.

Implementations

impl SubgroupPoint

pub const fn from_raw_unchecked(u: Fq, v: Fq) -> Self

Constructs an AffinePoint given u and v without checking that the point is on the curve or in the prime-order subgroup.

This should only be used for hard-coding constants (e.g. fixed generators); in all other cases, use SubgroupPoint::from_bytes instead.